Updated Hoedown.

+++ ext/hoedown/document.c
@@ -1356,7 +1356,7 @@ is_codefence(uint8_t *data, size_t size, size_t *width, uint8_t *chr)
 }
 
 /* expects single line, checks if it's a codefence and extracts language */
-static int
+static size_t
 parse_codefence(uint8_t *data, size_t size, hoedown_buffer *lang, size_t *width, uint8_t *chr)
 {
 	size_t i, w, lang_start;
